Split the search view apart
